refactor(math): 重构场数学节点以增强数值稳定性和类型安全
- 引入 FieldMath 工具类来处理有限值解析和正数验证
- 重命名 AttractorFieldBlendNode 为 Blend Vector Fields 并更新描述
- 重构标量和向量场二元运算节点以使用新的 FieldMath 组合操作
- 移除临时变量并改进字段采样实现以避免精度问题
- 添加 GraphFormatVersion V30 和相应的迁移逻辑处理字段版本控制
- 更新文档中的节点名称和描述以反映新的场数学实现
- 优化编辑器事件总线和列表序列节点的性能和类型检查

+++ src/main/java/com/nodecraft/nodesystem/execution/NodeExecutionCache.java
@@ -25,7 +25,7 @@ public final class NodeExecutionCache {
             return false;
         }
         Long lastExecuted = lastExecutedDirtyVersion.get(node.getId());
-        return lastExecuted != null && lastExecuted.longValue() == baseNode.getDirtyVersion();
+        return lastExecuted != null && lastExecuted == baseNode.getDirtyVersion();
     }
 
     public void invalidate(UUID nodeId) {
